Create transform files for each transform
