+++ kazaam.go
@@ -4,7 +4,6 @@ package kazaam
 import (
 	"encoding/json"
 	"fmt"
-	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
@@ -188,187 +187,10 @@ func (j *Kazaam) TransformJSONString(data string) (*simplejson.Json, error) {
 // return the transform function based on what's indicated in the operation spec
 func getTransform(specObj spec) transformFunc {
 	tform, ok := validSpecTypes[*specObj.Operation]
-	if ok == false {
+	if !ok {
 		return transformPass
 	}
 	return tform
-}
-
-// no op transform -- useful for testing/default behavior
-func transformPass(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	return data, nil
-}
-
-// simple transform to set default values in output json
-func transformDefault(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	for k, v := range *spec.Spec {
-		data.SetPath(strings.Split(k, "."), v)
-	}
-	return data, nil
-}
-
-func transformExtract(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	outPath, ok := (*spec.Spec)["path"]
-	if !ok {
-		return nil, &Error{ErrMsg: fmt.Sprintf("Unable to get path"), ErrType: SpecError}
-	}
-	outData, err := getJSONPath(data, outPath.(string), spec.Require)
-	return outData, err
-}
-
-func transformShift(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	outData := simplejson.New()
-	for k, v := range *spec.Spec {
-		array := true
-		outPath := strings.Split(k, ".")
-
-		var keyList []string
-
-		// check if `v` is a string or list and build a list of keys to evaluate
-		switch v.(type) {
-		case string:
-			keyList = append(keyList, v.(string))
-			array = false
-		case []interface{}:
-			for _, vItem := range v.([]interface{}) {
-				vItemStr, found := vItem.(string)
-				if !found {
-					return nil, &Error{ErrMsg: fmt.Sprintf("Warn: Unable to coerce element to json string: %v", vItem), ErrType: ParseError}
-				}
-				keyList = append(keyList, vItemStr)
-			}
-		default:
-			return nil, &Error{ErrMsg: fmt.Sprintf("Warn: Unknown type in message for key: %s", k), ErrType: ParseError}
-		}
-
-		// iterate over keys to evaluate
-		for _, v := range keyList {
-			var dataForV *simplejson.Json
-			var err error
-
-			// grab the data
-			if v == "$" {
-				dataForV = data
-			} else {
-				dataForV, err = getJSONPath(data, v, spec.Require)
-				if err != nil {
-					return nil, err
-				}
-			}
-
-			// if array flag set, encapsulate data
-			if array {
-				var intSlice = make([]interface{}, 1)
-				intSlice[0] = dataForV.Interface()
-				dataForV.SetPath(nil, intSlice)
-			}
-
-			outData.SetPath(outPath, dataForV.Interface())
-		}
-	}
-	return outData, nil
-}
-
-func transformConcat(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	sourceList, sourceOk := (*spec.Spec)["sources"]
-	if !sourceOk {
-		return nil, &Error{ErrMsg: fmt.Sprintf("Unable to get sources"), ErrType: SpecError}
-	}
-	targetPath, targetOk := (*spec.Spec)["targetPath"]
-	if !targetOk {
-		return nil, &Error{ErrMsg: fmt.Sprintf("Unable to get targetPath"), ErrType: SpecError}
-	}
-	delimiter, delimOk := (*spec.Spec)["delim"]
-	if !delimOk {
-		// missing delimiter.  default to blank
-		delimiter = ""
-	}
-
-	outString := ""
-	applyDelim := false
-	for _, vItem := range sourceList.([]interface{}) {
-		if applyDelim {
-			outString += delimiter.(string)
-		}
-		value, ok := vItem.(map[string]interface{})["value"]
-		if !ok {
-			path, ok := vItem.(map[string]interface{})["path"]
-			if ok {
-				valueNodePtr, err := getJSONPath(data, path.(string), spec.Require)
-				switch {
-				case err != nil && spec.Require == true:
-					return nil, &Error{ErrMsg: fmt.Sprintf("Path does not exist"), ErrType: RequireError}
-				case err != nil:
-					value = ""
-				default:
-					zed := (*valueNodePtr).Interface()
-					switch zed.(type) {
-					case []interface{}:
-						temp := ""
-						for _, item := range zed.([]interface{}) {
-							if item != nil {
-								temp += reflect.ValueOf(item).String()
-							}
-						}
-						value = temp
-					default:
-						value = reflect.ValueOf(zed).String()
-					}
-				}
-			} else {
-				return nil, &Error{ErrMsg: fmt.Sprintf("Error processing %v: must have either value or path specified", vItem), ErrType: SpecError}
-			}
-		}
-		outString += value.(string)
-
-		applyDelim = true
-	}
-
-	data.SetPath(strings.Split(targetPath.(string), "."), outString)
-
-	return data, nil
-}
-
-func transformCoalesce(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
-	if spec.Require == true {
-		return nil, &Error{ErrMsg: fmt.Sprintf("Invalid spec. Coalesce does not support \"require\""), ErrType: SpecError}
-	}
-	for k, v := range *spec.Spec {
-		outPath := strings.Split(k, ".")
-
-		var keyList []string
-
-		// check if `v` is a list and build a list of keys to evaluate
-		switch v.(type) {
-		case []interface{}:
-			for _, vItem := range v.([]interface{}) {
-				vItemStr, found := vItem.(string)
-				if !found {
-					return nil, &Error{ErrMsg: fmt.Sprintf("Warn: Unable to coerce element to json string: %v", vItem), ErrType: ParseError}
-				}
-				keyList = append(keyList, vItemStr)
-			}
-		default:
-			return nil, &Error{ErrMsg: fmt.Sprintf("Warn: Expected list in message for key: %s", k), ErrType: ParseError}
-		}
-
-		// iterate over keys to evaluate
-		for _, v := range keyList {
-			var dataForV *simplejson.Json
-			var err error
-
-			// grab the data
-			dataForV, err = getJSONPath(data, v, false)
-			if err != nil {
-				return nil, err
-			}
-			if dataForV.Interface() != nil {
-				data.SetPath(outPath, dataForV.Interface())
-				break
-			}
-		}
-	}
-	return data, nil
 }
 
 var jsonPathRe = regexp.MustCompile("([^\\[\\]]+)\\[([0-9\\*]+)\\]")
