+++ transform_concat_test.go
@@ -1,0 +1,201 @@
+package kazaam_test
+
+import (
+	"testing"
+
+	"github.com/qntfy/kazaam"
+)
+
+func TestConcat(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a.timestamp"}], "targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a":{"timestamp": 1481305274}}`
+
+	jsonOut := `{"a":{"output":"TEST,1481305274","timestamp":1481305274}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithRequireSources(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a":{"timestamp": 1481305274}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	_, err := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if err == nil {
+		t.Error("Source field is missing and should throw an error.")
+		t.FailNow()
+	}
+}
+
+func TestConcatWithRequireTargetPath(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a.timestamp"}], "delim": "," }}]`
+	jsonIn := `{"a":{"timestamp": 1481305274}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	_, err := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if err == nil {
+		t.Error("targetPath field is missing and should throw an error.")
+		t.FailNow()
+	}
+}
+
+func TestConcatWithRequireSimplePath(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "not.a.timestamp"}], "targetPath": "a.output", "delim": "," }, "require": true}]`
+	jsonIn := `{"a":{"timestamp": 1481305274}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	_, err := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if err == nil {
+		t.Error("Transform path does not exist in message and should throw an error")
+		t.FailNow()
+	}
+}
+
+func TestConcatWithReplaceSimplePath(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a.timestamp"}], "targetPath": "a.timestamp", "delim": "," }}]`
+	jsonIn := `{"a":{"timestamp": 1481305274}}`
+
+	jsonOut := `{"a":{"timestamp":"TEST,1481305274"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithNoDelimiter(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a.timestamp"}], "targetPath": "a.output" }}]`
+	jsonIn := `{"a":{"timestamp": "1481305274"}}`
+
+	jsonOut := `{"a":{"output":"TEST1481305274","timestamp":"1481305274"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithWildcard(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a[*].foo"}], "targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a":[{"foo": 0}, {"foo": 1}, {"foo": 1}, {"foo": 2}]}`
+
+	jsonOut := `{"a":{"output":"TEST,0112"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithWildcardNested(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a.b[*].foo"}], "targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a": {"b": [{"foo": 0}, {"foo": 1}, {"foo": 1}, {"foo": 2}]}}`
+
+	jsonOut := `{"a":{"b":[{"foo":0},{"foo":1},{"foo":1},{"foo":2}],"output":"TEST,0112"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithBadPath(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"path": "a[*].bar"}], "targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a":[{"foo": 0}, {"foo": 1}, {"foo": 1}, {"foo": 2}]}`
+
+	jsonOut := `{"a":{"output":"TEST,"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithBadSpec(t *testing.T) {
+	// Bad spec - "Path" should be "path"
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "TEST"}, {"Path": "a[*].bar"}], "targetPath": "a.timestamp", "delim": "," }}]`
+	jsonIn := `{"a":[{"foo": 0}, {"foo": 1}, {"foo": 1}, {"foo": 2}]}`
+
+	// bad path should cause the result to be blank
+	jsonOut := ""
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithMultiMulti(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"value": "BEGIN"}, {"path": "a[*].foo"}, {"value": "END"}], "targetPath": "a.output", "delim": "," }}]`
+	jsonIn := `{"a":[{"foo": 0}, {"foo": 1}, {"foo": 1}, {"foo": 2}]}`
+
+	jsonOut := `{"a":{"output":"BEGIN,0112,END"}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
+
+func TestConcatWithLargeNumbers(t *testing.T) {
+	spec := `[{"operation": "concat", "spec": {"sources": [{"path": "a.timestamp"}], "targetPath": "a.output" }}]`
+	jsonIn := `{"a":{"timestamp": 1481305274100000000000000000000}}`
+
+	jsonOut := `{"a":{"output":"1481305274100000000000000000000","timestamp":1481305274100000000000000000000}}`
+
+	kazaamTransform, _ := kazaam.NewKazaam(spec)
+	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(jsonIn)
+
+	if kazaamOut != jsonOut {
+		t.Error("Transformed data does not match expectation.")
+		t.Log("Expected: ", jsonOut)
+		t.Log("Actual:   ", kazaamOut)
+		t.FailNow()
+	}
+}
